doc: synonym propeq generation
