+++ src/composer/compose_predicate.rs
@@ -27,95 +27,7 @@ use crate::composer::postulate::AgdaStructure::FunctionDef;
 use crate::wordnet::interface::{get_meanings, init_wordnet};
 use crate::wordnet::wordnode::Wordnode;
 use crate::composer::ast::*;
-
-
-
-pub enum SynsetStrategy {
-    Ignore, BestMatch, AllMeanings
-}
-
-
-
-pub enum SynsetRelevancyStrategy {
-    Ignore, Relevant, All
-}
-
-
-pub fn handle_synonyms(property: &str, f: &mut AgdaFile) {
-
-    let SYNSTRAT: SynsetStrategy = SynsetStrategy::AllMeanings;
-    let SYNRELEVANT: SynsetRelevancyStrategy = SynsetRelevancyStrategy::Relevant;
-
-    if let SynsetStrategy::Ignore = SYNSTRAT { return; }
-
-    let wordnet_neighbours = get_meanings(property).unwrap_or_default();
-
-    let words_in_existence_snapshot = {
-        let words_in_existence = WORDS_IN_EXISTENCE.lock().unwrap();
-        words_in_existence.clone()
-    };
-
-    for wordnode in wordnet_neighbours {
-        for synonym in wordnode.synonyms {
-
-            /* Often, it's possible to prune the search by only looking for other words used in the sentence */
-            if matches!(SYNRELEVANT, SynsetRelevancyStrategy::Relevant) && !words_in_existence_snapshot.contains(&synonym) {
-                continue;
-            } else {
-                let is_property = convert_case(format!("is_{}", property).as_str(), CaseStyle::CamelCase);
-                let is_synonym = convert_case(format!("is_{}", synonym).as_str(), CaseStyle::CamelCase);
-
-                println!("Adding : {property} ≡ {synonym}");
-                let equality_identifier: String = format!("{}_syn_{}", property, synonym);
-                f.insert_postulate(PostulateEntry(
-                    equality_identifier.clone(),
-                    *τPropEq!(τSimp!(is_property.clone()), τSimp!(is_synonym.clone())),
-                ));
-
-                let ast = astLambda!(
-                    String::from("e"),
-                    astLambda!(
-                        String::from("m"),
-                        astApply!(
-                            astApply!(
-                                astTerm!(String::from("subst")),
-                                astLambda!(
-                                    String::from("X"),
-                                    astApply!(astTerm!(String::from("X")), astTerm!(String::from("e")))
-                                )
-                            ),
-                            astApply!(
-                                astTerm!(equality_identifier.clone()),
-                                astTerm!(String::from("m"))
-                            )
-                        )
-                    )
-                );
-                let type_header =
-                    τDepFunc!(
-                    "e".parse().unwrap(), τSimp!("Entity".parse().unwrap()),
-                    τFunc!(τApp!(τSimp!(is_property), τSimp!("e".parse().unwrap())), τApp!(τSimp!(is_synonym),  τSimp!("e".parse().unwrap())))
-                );
-
-                f.insert_definition(FunctionDef(FunctionDefinition {
-                    function_name: (*format!("{}_syn_{}_pointwise", property, synonym)).parse().unwrap(),
-                    function_type: *type_header,
-                    function_body: *ast,
-                }));
-
-                /* Derive a Pointwise Equality Function from this Equality */
-                // MortalPerishablePointwise : (e : Entity) → isMortal e → isPerishable e
-                // MortalPerishablePointwise e m = subst (λ X → X e) eqMortalPerishable m
-
-                // make an ast here following this general form:
-                // {property}_{synonym}_pointwise = lambda e -> lambda m -> subst ( lambda X to X e) {property}_syn_{synonym}
-
-
-            }
-        }
-        if let SynsetStrategy::BestMatch = SYNSTRAT { return; }
-    }
-}
+use crate::composer::synonym_handler::handle_synonyms;
